END-284 Help texts are not displayed if JavaScript is not enabled
Since only the headings of the help texts are displayed initially, and
the user has to click on them to read the entire text (and for this
JavaScript is required), we hide the help texts for non-JavaScript
users. Another solution would be to always display all text, but that
seems less good.